% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_Abby.R
\name{run_app_Abby}
\alias{run_app_Abby}
\title{Run the standalone version}
\usage{
run_app_Abby(options = list())
}
\arguments{
\item{options}{Named options that should be passed to the runApp call.}
}
\value{
Open browser
}
\description{
Run the standalone version
}
