% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABI_Helminth.R
\name{ABI_Helminth}
\alias{ABI_Helminth}
\title{A tool for helminth species delimitation at various taxonomic levels}
\usage{
ABI_Helminth(distance = 0, group = "NAS", marker = "18S rRNA")
}
\arguments{
\item{distance}{Number >= 0 (Number should be 0.000 - 0.735)}

\item{group}{group of Helminth ("NAS","NS","NT","TR","CE")
"NAS" is "Nematode (Ascaridida and Spirurida)"
"NS"  is "Nematode (Strongylida)")
"NT"  is "Nematode (Trichocephalida)"
"TR"  is "Trematode"
"CE"  is "Cestode"}

\item{marker}{Helminth Genetic Markers
("18S rRNA","28S rRNA","ITS1","ITS2","COI","COII","cytB","NAD1","12S rRNA","16S rRNA")}
}
\value{
Plot of ggplot
}
\description{
A tool for helminth species delimitation at various taxonomic levels
}
