% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_ABI.R
\name{run_app_ABI}
\alias{run_app_ABI}
\title{Run the standalone version}
\usage{
run_app_ABI(options = list())
}
\arguments{
\item{options}{Named options that should be passed to the runApp call.}
}
\value{
Open browser
}
\description{
Run the standalone version
}
\examples{
run_app_ABI()
}
